IP 0.0.0.0 breaks loading on windows
Hi!

I was looking for an nwb starter for electron and I found yours. Nice job packaging things together so cleanly.

While testing it on my windows setup I found that the 0.0.0.0 ip address on the main.js for devevelopment setup wasn't being resolved to localhost as the dev server was setup. This small change made it work as intended on windows.